+++ tests/sdk/test_rest_client.py
@@ -24,6 +24,8 @@ from .utils import mock_requests
 from .utils import mock_requests_responses
 from .utils import mock_response
 
+from .. import datastore
+
 CONFIG = {
     "url": "http://foo.com",
     "insecure": False,
@@ -113,3 +115,40 @@ def test_add_wrong_url(mocker):
     with pytest.raises(exceptions.BadConfiguration) as e:
         test_client.login("test_client", "hehe")
     assert "Make sure that given url" in e.value.args[0]
+
+
+def test_list_paginated(mocker):
+    asset_name = "traintuple"
+    items = [datastore.TRAINTUPLE, datastore.TRAINTUPLE]
+    responses = [
+        mock_response(response={
+                "count": len(items),
+                "next": "http://foo.com/?page=2",
+                "previous": None,
+                "results": items[:1],
+            }, status=200),
+        mock_response(response={
+                "count": len(items),
+                "next": None,
+                "previous": "http://foo.com/?page=1",
+                "results": items[1:],
+            }, status=200),
+    ]
+    m_get = mock_requests_responses(mocker, "get", responses)
+    asset = _client_from_config(CONFIG).list(asset_name)
+    assert len(asset) == len(items)
+    assert len(m_get.call_args_list) == 2
+
+
+def test_list_not_paginated(mocker):
+    asset_name = "traintuple"
+    items = [datastore.TRAINTUPLE, datastore.TRAINTUPLE]
+    m_get = mock_requests(mocker, "get", response={
+                "count": len(items),
+                "next": "http://foo.com/?page=2",
+                "previous": None,
+                "results": items[:1],
+            }, status=200)
+    asset = _client_from_config(CONFIG).list(asset_name, paginated=False)
+    assert len(asset) != len(items)
+    assert len(m_get.call_args_list) == 1